+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 RTLD 0.6.0 - TBD
 ----------------
+* **rtld.core**
+  * New module `rtld.core.conv`
+  * New module `rtld.core.errno`
 * **rtld.data**
   * New package `rtld.data` for data formats implementations. Currently it contains JSON parser (`rtld.data.json`) and Varint encoder/decoder (`rtld.data.varint`)
 * **rtld.container**